INCOMPAT: cleanup directive, jedi cache and mist updates.
This cleanup inline directive to store a string instead of a list of
strings, which is simpler and more compact,
we also make sure the value does not contain backticks to catch more
syntax errors.

This leads to a few PR upstream and a few skipped object

We also add some caching to jedi, so speed a bit repetitive build.
The cache is probalistically invalidated every 1h