This is an attempt as writing a cbor serializer.
Unlike the json serializer it should be much more compact.
This remove 1 level of Json nesting, and the type strings get simplified
to CBor tags.

The drawback is that tags needs to be assigned, which is a manual
process.

From a small experiment for object with lots of text, the reduction is
minimal. For object with minimal text but lot of structure, i've seen
reduction generally arround 2x-3x and up to 6x