Add option to regenerate documentation only for some items.
When debugging or iterating on the docs it should be interesting to
be able to update the intermediate bundle with only some elements
For example because you are updating the docs.

This adds a flag --only , that can be passed multiple time at the CLI,
and will ask the gen step to only collect documentation for those
objects, and update the doc bundle instead of regenerating it.

It is likely imperfect and  hence why it's only a cli flag.

I'm not happy withe the name --only. Maybe --update ?