Misc Ascii and browser fixes.
This move some of the coloration outside of the templating engine, this
should let us add an option to make colors optionals, and should make
testing easier, as now we can have expected colorless outputs and test
the codebase against it.

This also add some fixes to the urwid frontend – which still crashes in
a number of places, so I'm wondering if we shoudl rewrite it using
textual. One of the other current limitation with urwind is the lack of
ability to reflow paragraph is they contain multiple entries.