Don't store a file for dynamically generated item.
We try to detect where a item was define, sometime methods/objects are
dynamically defined.

When this is the case, we get back a file that contain `<string>` in
it's name. When that's the case we store `None` instead.

I'm wondering if we should store something else, as `None` also mean we
were not able to find what the source was...