One thing I'd like to do is do the full rendering with myst.js,
this will make it much more easier to make a jupyterLab extension as once we have a JS renderer, we can "just" make a JupyterLab extension that use the same components.

I dont' know much about how to do that, but here is an attempt that try to do it:

- it add the the "frontend" folder, if you run `npm install` (once) and then `npm run build` from within it, it will generate the a `index.<hash>.js` file in `papyri/app/js/...`.

As I don't know how to not have it have a `hash`, this adds new handlers that try to guess the name of the file and serve it.

This let you in a template to call `render(dom-id, json)`, and ti wil use myst to render that on the given node.

The idea is to give myst the root node, but that does not work so far as some of our nodes are not the same.

we can go two routes (or a mix of both):
 - write custom rendered for myst,
 - massage our nodes to look like what myst is expecting.