Partial default value fix.
This is a partial fix for signature default values.

Currently we use an ugly way to display signature by converting "back"
into a Signature object, but we do lose some information as we have to
turn most of these into strings.

It does only affects rendering though, so I think this is ok for now