Fix out of order section rendering.
Closes #344

This is a bit of an ugly implementation but it will do for now.

As we are storing the content in a dict the order can be undone at
serialisation/deserialisation.

We had a list with the order of sections, but it was not being kept in
sync. In order to not change the API to much, I inserted a proxy that
mimic an ordered dict.

I also remove the readthedocs package as we never build/install it.