Do not allow Newline in inline nodes.
When a literal span two lines, as we directly extract the bytes offsets
from tree sitter, we do end up including the newline.

This make it slightly more complicated to fix in the rendering as we now
need to consider newline separately in the wrapping logic.

With this we normalize newline to spaces as if the text was all written
on the same line.