Fix rich rendering of panels
We use a table with a single column, single header and/or single row that
contain the renderables.

The border color as well as the icon in title depends on the kind of
admonition (maybe we want to limit the kind of admonitions.

We cannot render only a header, so when the directive has no content, we
render only the body.